now parsing both ROLLBACK and ABORT

+++ tpcc_replay_run.sh
@@ -7,7 +7,7 @@ PERMANENT_LOG_CSV_FPATH="$DBLAB_POSTGRES_DPATH/tpcc_log.csv"
 SKIP_BUILD=""
 
 # generate log file
-if true; then
+if false; then
   # this if block creates tpcc_log.csv
   SKIP_BUILD=$SKIP_BUILD ./run.sh tpcc-c-l
   SKIP_BUILD="1"
@@ -24,8 +24,10 @@ if true; then
   SKIP_BUILD="1" # if it was already 1, this is harmless
 fi
 
-# clean log file before running the replay
-rm "$(readlink $GEN_LOG_CSV_SPATH)"
+if true; then
+  # clean log file before running the replay so that we can track that as well
+  rm "$(readlink $GEN_LOG_CSV_SPATH)"
 
-# run replay
-SKIP_BUILD=$SKIP_BUILD ./run.sh replay # then, run a replay
+  # run replay
+  SKIP_BUILD=$SKIP_BUILD ./run.sh replay # then, run a replay
+fi
